feat: update dependencies and improve batch preparation
- Changed the branch of the simplecv dependency from 'ingest-all' to 'main'.
- Added new dependencies: gradio, huggingface_hub, tomlkit, audioop-lts, and pydub.
- Modified the data type of seg_map in final_demo.py from uint16 to int32 for better compatibility.
- Enhanced the prepare_batch function in prepare_batch.py by introducing a TypedDict for better type safety and clarity.
- Updated function signatures and added type hints for improved code readability and maintainability.

+++ packages/sam3d_body/src/sam3d_body/api/final_demo.py
@@ -164,7 +164,7 @@ def visualize_sample(
 
     # Build per-pixel maps (SEG_CLASS_OFFSET = background). Also build RGBA overlay with transparent background.
     h, w = bgr_image.shape[:2]
-    seg_map: Int[ndarray, "h w"] = np.full((h, w), SEG_CLASS_OFFSET, dtype=np.uint16)
+    seg_map: Int[ndarray, "h w"] = np.full((h, w), SEG_CLASS_OFFSET, dtype=np.int32)
     seg_overlay: UInt8[ndarray, "h w 4"] = np.zeros((h, w, 4), dtype=np.uint8)
 
     mesh_root_path: Path = parent_log_path / "pred"
